+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### Added
 - support for [polylith](https://polylith.gitbook.io/polylith/) layout 
+- support envsubst style semantics for [:target-image :image-name]
 
 ### Changed
 - default WORKING_DIR changed to `/home/app` (was previously `/`)